backend/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
aliyun-python-sdk-core==2.13.36
amqp==5.1.1
anyio==3.7.0
asgiref==3.6.0
astor==0.8.1
async-timeout==4.0.2
attrs==22.2.0
autobahn==23.1.2
Automat==22.10.0
Babel==2.12.1
bce-python-sdk==0.8.83
bcrypt==4.0.1
billiard==3.6.4.0
blinker==1.6.2
celery==5.2.7
certifi @ file:///C:/b/abs_85o_6fm0se/croot/certifi_1671487778835/work/certifi
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
chardet==5.1.0
charset-normalizer==3.1.0
ci-info==0.3.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
colorlog==6.7.0
configobj==5.0.8
configparser==5.3.0
constantly==15.1.0
contourpy==1.1.0
coreapi==2.3.3
coreschema==0.0.4
cron-descriptor==1.2.35
cryptography==40.0.1
cycler==0.11.0
daphne==4.0.0
datasets==2.13.0
decorator==5.1.1
dill==0.3.4
Django==4.2
django-celery-beat==2.5.0
django-celery-results==2.5.0
django-comment-migrate==0.1.7
django-cors-headers==3.14.0
django-filter==23.1
django-ranged-response==0.2.0
django-redis==5.2.0
django-simple-captcha==0.5.17
django-timezone-field==5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
dnspython==2.3.0
drf-yasg==1.21.5
etelemetry==0.3.0
eventlet==0.33.3
exceptiongroup==1.1.1
fastapi==0.97.0
filelock==3.12.2
fitz==0.0.1.dev2
Flask==2.3.2
Flask-Babel==2.0.0
fonttools==4.40.0
frozenlist==1.3.3
fsspec==2023.6.0
future==0.18.3
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.2
httplib2==0.22.0
httpx==0.24.1
huggingface-hub==0.15.1
hyperlink==21.0.0
idna==3.4
importlib-metadata==6.7.0
importlib-resources==5.12.0
incremental==22.10.0
inflection==0.5.1
isodate==0.6.1
itsdangerous==2.1.2
itypes==1.2.0
jieba==0.42.1
Jinja2==3.1.2
jmespath==0.10.0
joblib==1.2.0
jpush==3.3.9
kiwisolver==1.4.4
kombu==5.2.4
LAC==2.1.2
looseversion==1.2.0
lxml==4.9.2
markdown-it-py==3.0.0
MarkupSafe==2.1.2
matplotlib==3.7.1
mdurl==0.1.2
msgpack==1.0.5
multidict==6.0.4
multiprocess==0.70.12.2
mysqlclient==2.1.1
networkx==3.1
nibabel==5.1.0
nipype==1.8.6
numpy==1.25.0
opt-einsum==3.3.0
packaging==23.0
paddle-bfloat==0.1.7
paddle2onnx==1.0.6
paddlefsl==1.1.0
paddlenlp==2.5.2
paddlepaddle==2.4.2
pandas==2.0.2
paramiko==3.1.0
pathlib==1.0.1
Pillow==9.5.0
prompt-toolkit==3.0.38
protobuf==3.20.0
prov==2.0.0
psutil==5.9.4
pyarrow==12.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.17
pycryptodomex==3.17
pydantic==1.10.9
pydot==1.4.2
Pygments==2.15.1
PyJWT==2.6.0
PyMuPDF==1.22.5
PyNaCl==1.5.0
pyOpenSSL==23.1.1
pyparsing==3.0.9
python-crontab==2.7.1
python-dateutil==2.8.2
pytz==2023.3
pywin32==306
pyxnat==1.5
PyYAML==6.0
rdflib==6.3.2
redis==4.5.4
requests==2.28.2
rich==13.4.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
scikit-learn==1.2.2
scipy==1.10.1
sentencepiece==0.1.99
seqeval==1.2.2
service-identity==21.1.0
simplejson==3.19.1
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.3
starlette==0.27.0
tencentcloud-sdk-python==3.0.870
threadpoolctl==3.1.0
tqdm==4.65.0
traits==6.3.2
Twisted==22.10.0
twisted-iocpsupport==1.0.3
txaio==23.1.1
typer==0.9.0
typing_extensions==4.5.0
tzdata==2023.3
ua-parser==0.16.1
uritemplate==4.1.1
urllib3==1.26.15
user-agents==2.2.0
uvicorn==0.22.0
vine==5.0.0
visualdl==2.4.2
wcwidth==0.2.6
Werkzeug==2.3.6
wincertstore==0.2
xlrd==2.0.1
xlwt==1.3.0
xxhash==3.2.0
yarl==1.9.2
zipp==3.15.0
zope.interface==6.0
